preco now is float

+++ src/main/java/Lista02_HerancaPolimorfismo/Q8/Veiculo.java
@@ -5,14 +5,14 @@ public abstract class Veiculo {
     protected String modelo;
     protected float preco;
 
-    public void setPreco(int preco) {
+    public void setPreco(float preco) {
         if (preco < 0){
             throw new IllegalArgumentException("Preço não pode ser negativo");
         }
         this.preco = preco;
     }
 
-    public Veiculo(String marca, String modelo, int preco) {
+    public Veiculo(String marca, String modelo, float preco) {
         this.marca = marca;
         this.modelo = modelo;
         this.setPreco(preco);
